Android - Upgrade RN to 0.78.3 and apply new architecture
- Followed RN Upgrade helper instructions to migrate from 0.70.0 to 0.78.3: https://react-native-community.github.io/upgrade-helper/?from=0.70.0&to=0.78.3&package=com.example&name=example

Related to YD-6861
