+++ android/build.gradle
@@ -7,16 +7,13 @@ def safeExtGet(prop, fallback) {
 android {
     namespace "com.yoti.reactnative.docscan"
 
-    compileSdk safeExtGet('compileSdkVersion', 34)
+    compileSdk safeExtGet('compileSdkVersion', 35)
 
     defaultConfig {
-        minSdkVersion safeExtGet('minSdkVersion', 21)
-        targetSdkVersion safeExtGet('targetSdkVersion', 34)
+        minSdkVersion safeExtGet('minSdkVersion', 24)
+        targetSdkVersion safeExtGet('targetSdkVersion', 35)
         versionCode 602
         versionName "5.0.2"
-        ndk {
-            abiFilters "armeabi-v7a", "x86"
-        }
     }
     buildTypes {
         debug {
