--- example/android/app/src/main/java/com/example/MainActivity.java
@@ -1,10 +1,0 @@
-package com.example;
-
-import com.facebook.react.ReactActivity;
-
-public class MainActivity extends ReactActivity {
-  @Override
-  protected String getMainComponentName() {
-    return "example";
-  }
-}
