+++ example/android/build.gradle
@@ -1,34 +1,40 @@
 buildscript {
     ext {
-        minSdkVersion = 21
-        compileSdkVersion = 34
-        targetSdkVersion = 34
-        yotiSdkVersion = "4.1.0"
+        buildToolsVersion = "35.0.0"
+        minSdkVersion = 24
+        compileSdkVersion = 35
+        targetSdkVersion = 35
+        ndkVersion = "27.1.12297006"
+        kotlinVersion = "2.0.21"
+        yotiSdkVersion = "4.2.1"
     }
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath('com.android.tools.build:gradle:8.8.0')
+        classpath("com.android.tools.build:gradle")
+        classpath("com.facebook.react:react-native-gradle-plugin")
+        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin")
     }
 }
+
+apply plugin: "com.facebook.react.rootproject"
 
 allprojects {
     repositories {
         mavenLocal()
+        mavenCentral()
+        google()
         maven {
             url("$rootDir/../node_modules/react-native/android")
         }
         maven {
             url("$rootDir/../node_modules/jsc-android/dist")
         }
-        google()
-        jcenter()
         maven {
             url 'https://jitpack.io'
         }
-        mavenCentral()
         maven {
             url 'https://maven.regulaforensics.com/RegulaDocumentReader'
         }
