+++ example/android/gradle.properties
@@ -20,3 +20,8 @@ org.gradle.jvmargs=-Xmx2048m
 
 android.useAndroidX=true
 android.enableJetifier=true
+newArchEnabled=true
+ 
+# Use this property to enable or disable the Hermes JS engine.
+# If set to false, you will be using JSC instead.
+hermesEnabled=true
