+++ example/package.json
@@ -5,17 +5,36 @@
     "scripts": {
       "android": "react-native run-android",
       "ios": "react-native run-ios",
-      "start": "react-native start"
+      "start": "react-native start",
+      "lint": "eslint .",
+      "test": "jest"
     },
     "dependencies": {
-      "react": "^18.0.0",
-      "react-native": "^0.70.0",
+      "react": "19.0.0",
+      "react-native": "0.78.3",
       "@getyoti/yoti-doc-scan-react-native": "file:.."
     },
     "devDependencies": {
-      "@babel/core": "^7.20.0",
-      "@babel/runtime": "^7.20.0",
-      "metro-react-native-babel-preset": "^0.70.0",
-      "react-test-renderer": "^18.0.0"
+      "@babel/core": "^7.25.2",
+      "@babel/preset-env": "^7.25.3",
+      "@babel/runtime": "^7.25.0",
+      "@react-native-community/cli": "15.0.1",
+      "@react-native-community/cli-platform-android": "15.0.1",
+      "@react-native-community/cli-platform-ios": "15.0.1",
+      "@react-native/babel-preset": "0.78.3",
+      "@react-native/eslint-config": "0.78.3",
+      "@react-native/metro-config": "0.78.3",
+      "@react-native/typescript-config": "0.78.3",
+      "@types/jest": "^29.5.13",
+      "@types/react": "^19.0.0",
+      "@types/react-test-renderer": "^19.0.0",
+      "eslint": "^8.19.0",
+      "jest": "^29.6.3",
+      "prettier": "2.8.8",
+      "react-test-renderer": "19.0.0",
+      "typescript": "5.0.4"
+    },
+    "engines": {
+      "node": ">=18"
     }
 }
